requirement.txt
generativeai==0.0.1
google-ai-generativelanguage==0.6.15
pymongo==4.15.5
chromadb==1.3.5
requests==2.32.3